+++ protected/humhub/tests/codeception/_support/FunctionalTester.php
@@ -4,6 +4,8 @@ use humhub\modules\friendship\models\Friendship;
 use humhub\modules\space\models\Space;
 use tests\codeception\_pages\LoginPage;
 use humhub\modules\user\models\User;
+use humhub\modules\user\models\GroupPermission;
+use yii\helpers\Url;
 
 /**
  * Inherited Methods
@@ -36,19 +38,100 @@ class FunctionalTester extends \Codeception\Actor
         $this->see('Administration');
     }
 
-    function setGroupPermission($groupId, $permission, $state = 1)
+    public function setGroupPermission($groupId, $permission, $state = 1)
     {
-        $groupPermission = new humhub\modules\user\models\GroupPermission();
-        $groupPermission->permission_id = $permission->id;
-        $groupPermission->group_id = $groupId;
-        $groupPermission->module_id = $permission->moduleId;
-        $groupPermission->class = $permission->className();
-        $groupPermission->state = $state;
-        $groupPermission->save();
+        if(is_string($permission)) {
+            $permission = Yii::createObject(['class' => $permission]);
+        }
+
+        (new GroupPermission([
+            'permission_id' => $permission->id,
+            'group_id' => $groupId,
+            'module_id' => $permission->moduleId,
+            'class' => get_class($permission),
+            'state' => $state
+        ]))->save();
+
         \Yii::$app->user->getPermissionManager()->clear();
     }
 
-    public function amUser($user = null, $password = null, $logout = false)
+    public function assertSpaceAccessStatus($userGroup, $status, $path, $params = [], $post = false)
+    {
+        $space = $this->loginBySpaceUserGroup($userGroup, $path, $params, $post);
+        $this->seeResponseCodeIs($status);
+        $this->logout();
+        return $space;
+    }
+
+    public function assertSpaceAccessTrue($userGroup, $path, $params = [], $post = false)
+    {
+        $space = $this->loginBySpaceUserGroup($userGroup, $path, $params, $post);
+        $this->seeSuccessResponseCode();
+        $this->logout();
+        return $space;
+    }
+
+    public function seeSuccessResponseCode()
+    {
+        $this->seeResponseCodeIsBetween(200, 308);
+    }
+
+    public function assertSpaceAccessFalse($userGroup, $path, $params = [], $post = false)
+    {
+        $space = $this->loginBySpaceUserGroup($userGroup, $path, $params, $post);
+        $this->dontSeeResponseCodeIs(200);
+        $this->logout();
+        return $space;
+    }
+
+    /**
+     * This utility function finds a space membership relation of the given $userGroup and logs in the user and
+     * also access the space and returns the related space model.
+     *
+     * @param $userGroup
+     * @return Space|null
+     */
+    public function loginBySpaceUserGroup($userGroup, $path = null, $params = [], $post = false)
+    {
+        $spaceId = null;
+        $user = null;
+        switch($userGroup) {
+            case 'root':
+                $spaceId = 2;
+                $user = 'Admin';
+                break;
+            case Space::USERGROUP_OWNER:
+                $spaceId = 2;
+                $user = 'User1';
+                break;
+            case Space::USERGROUP_ADMIN:
+                $spaceId = 4;
+                $user = 'User1';
+                break;
+            case Space::USERGROUP_MODERATOR:
+                $spaceId = 3;
+                $user = 'User2';
+                break;
+            case Space::USERGROUP_MEMBER:
+                $spaceId = 3;
+                $user = 'User1';
+                break;
+            case Space::USERGROUP_GUEST:
+                $spaceId = 1;
+                $user = 'User1';
+                break;
+        }
+
+        $space = Space::findOne(['id' => $spaceId]);
+
+
+        $this->amUser($user);
+        $this->amOnSpace($space, $path, $params, $post);
+        return $space;
+    }
+
+
+    public function amUser($user = null, $password = '123qwe', $logout = false)
     {
         if ($logout) {
             $this->logout();
@@ -57,6 +140,9 @@ class FunctionalTester extends \Codeception\Actor
         if ($user == null) {
             $this->amUser1();
         } else {
+            if(strtolower($user) == 'admin') {
+                $password = 'test';
+            }
             LoginPage::openBy($this)->login($user, $password);
             tests\codeception\_pages\DashboardPage::openBy($this);
             $this->see('Dashboard');
@@ -113,24 +199,24 @@ class FunctionalTester extends \Codeception\Actor
         $user->profile->save();
     }
 
-    public function amOnSpace1($path = '/space/space', $params = [])
+    public function amOnSpace1($path = '/space/space', $params = [], $post = false)
     {
-        $this->amOnSpace(1, $path, $params);
+        $this->amOnSpace(1, $path, $params, $post);
     }
 
-    public function amOnSpace2($path = '/space/space', $params = [])
+    public function amOnSpace2($path = '/space/space', $params = [], $post = false)
     {
-        $this->amOnSpace(2, $path, $params);
+        $this->amOnSpace(2, $path, $params, $post);
     }
 
-    public function amOnSpace3($path = '/space/space', $params = [])
+    public function amOnSpace3($path = '/space/space', $params = [], $post = false)
     {
-        $this->amOnSpace(3, $path, $params);
+        $this->amOnSpace(3, $path, $params, $post);
     }
 
-    public function amOnSpace4($path = '/space/space', $params = [])
+    public function amOnSpace4($path = '/space/space', $params = [], $post = false)
     {
-        $this->amOnSpace(4, $path, $params);
+        $this->amOnSpace(4, $path, $params, $post);
     }
 
     public $spaces = [
@@ -140,19 +226,32 @@ class FunctionalTester extends \Codeception\Actor
         '5396d499-20d6-4233-800b-c6c86e5fa34d',
     ];
 
-    public function amOnSpace($guid, $path = '/space/space', $params = [])
+    public function amOnSpace($guid, $path = '/space/space', $params = [], $post = false)
     {
+        if(is_bool($params)) {
+            $post = $params;
+            $params = [];
+        }
+
         if(!$path) {
             $path = '/space/space';
         }
 
         if(is_int($guid)) {
             $guid = $this->spaces[--$guid];
+        } else if($guid instanceof Space) {
+            $guid = $guid->guid;
         }
 
-        $params['sguid'] = $guid;
+        $params['cguid'] = $guid;
 
-        $this->amOnRoute($path, $params);
+        if($post) {
+            $route =  array_merge([$path], $params);
+            $this->sendAjaxPostRequest(Url::toRoute($route), (is_array($post) ? $post : []));
+        } else {
+            $this->amOnRoute($path, $params);
+        }
+
     }
 
     public function enableModule($guid, $moduleId)